Bug:58045 - Tab overflow
http://bugzilla.zimbra.com/show_bug.cgi?id=58045

+++ ZimbraWebClient/WebRoot/js/ajax/dwt/widgets/DwtMenu.js
@@ -917,7 +917,7 @@ function(child) {
 			this._table.rows[0].deleteCell(Dwt.getCellIndex(cell));
 		} else {
 			var el = child.getHtmlElement();
-			if (el)
+			if (el && el.parentNode.parentNode.rowIndex > -1)// Make sure that the element exists in the table
 				this._table.deleteRow(el.parentNode.parentNode.rowIndex);
 		}
 	}
